[Misc]] Move processing context to multimodal directory
