+++ tests/multimodal/test_processing.py
@@ -8,11 +8,11 @@ import numpy as np
 import pytest
 
 from vllm.config import ModelConfig
-from vllm.inputs import InputProcessingContext
 from vllm.multimodal import MULTIMODAL_REGISTRY
 # yapf conflicts with isort for this block
 # yapf: disable
-from vllm.multimodal.processing import (PlaceholderFeaturesInfo,
+from vllm.multimodal.processing import (InputProcessingContext,
+                                        PlaceholderFeaturesInfo,
                                         PromptIndexTargets, PromptInsertion,
                                         PromptReplacement, apply_text_matches,
                                         apply_token_matches,
